+++ test/test_service.py
@@ -93,16 +93,6 @@ class ServiceTest(TestCase):
         env = service.environment
         self.assertIsInstance(env, Environment)
 
-    def test_parse_config_files(self):
-        service = s.Service()
-        env = service.environment
-        env.config_file_paths.append('test/')
-        tornado.options.define('test', default='default')
-        service.parse_config_files()
-
-        from tornado.options import options
-        self.assertEqual('filevalue', options.test)
-
     def test_logging_initialization(self):
         service = s.Service()
         env = service.environment
@@ -219,3 +209,22 @@ def test_system_exit_after_showing_config(option_name):
         service = s.Service()
         with pytest.raises(SystemExit):
             service.get_app()
+
+
+@pytest.fixture
+def test_env_var(monkeypatch):
+    monkeypatch.setenv("TEST", "envvalue")
+
+
+def test_config_parsing(test_env_var):
+    service = s.Service()
+    env = service.environment
+    env.config_file_paths.append('test/')
+    tornado.options.define('test', default='default')
+    service.parse_config_files()
+
+    from tornado.options import options
+    assert 'filevalue' == options.test
+
+    service.parse_environment_variables(options)
+    assert 'envvalue' == options.test